Version Partitional v0.03
frontend present, demands small refactor
Business logic exists, better to refactor.
Services demands verification.( verify data and murkup on page)
DAO finished, tested(tests are not looking good, but DAO works OK)
loggers set up but not present everywhere(better to refactor)
internationalization is absent
storing passwords and cookies in MD5 required
error page is not designed and set up!
check Filtres one more time and remove root/ from path!
check multithreding

+++ src/main/java/com/fluffypets/DAO/orders/OrderDAO.java
@@ -6,5 +6,5 @@ import com.fluffypets.MVC.model.Order;
 import java.util.List;
 
 public interface OrderDAO extends GenericDAO<Order> {
-    List<Order> getMyOrders(Long userId);
+    List<Order> getMyOrders(Integer userId);
 }
