+++ src/main/java/com/fluffypets/servicies/UserDataService.java
@@ -4,7 +4,7 @@ import com.fluffypets.MVC.model.UserData;
 
 public interface UserDataService {
 
-    UserData get(Long userId);
+    UserData get(Integer userId);
 
     UserData create(UserData userData);
